services/imgs/requirements.txt
fbrecog==1.6
Flask==0.12.2
Flask-Script==2.0.5
flask-sqlalchemy==2.3.2
psycopg2==2.7.3.1
Flask-Testing==0.6.2
gunicorn==19.7.1